Android overflow fix & prop not applying fix
This fix will make fix the overflow issue that was seen on Android SDK 26+, where in the checkbox in checked mode, fills the entire screen. This fix also contains a patch that helps in avoiding scenarios where the checkbox doesn't changes it's checked state, (as encountered on iPhone X running iOS 12).