Fix for Checkbox not updating visual state, on iOS
This fix addresses the issue when on devices like iPhone X running iOS 12, when we try to check the checkbox initially, it shows the ripple, and gives it's new state in code but visually it doesn't changes it's state. That is, it looks like it's not checked (whereas internally it is). So, this fix shall fix this inconsistency between internal state and visual state of this component.